Add tests of current text output handling
