+++ src/NUnitFramework/tests/nunit.framework.tests-2.0.csproj
@@ -282,6 +282,7 @@
     <Compile Include="TestUtilities\TestFinder.cs" />
     <Compile Include="Constraints\ExactCountConstraintTests.cs" />
     <Compile Include="TestUtilities\UniqueValues.cs" />
+    <Compile Include="TextOutputTests.cs" />
     <Compile Include="ThrowsTests.cs" />
   </ItemGroup>
   <ItemGroup>
